omissions and explanations
I took out the screenshot because I was unsure of its identity. also, I forgot to add a branch earlier so my commits are in the main branch
I hope that is not too bad as this message confirms that I do know how to navigate a main branch and its feature branches its just a matter  of me forgetting
thanks.